+++ trunk/decoder_plugins/mp3/mp3.c
@@ -23,6 +23,7 @@
 
 #include <unistd.h>
 #include <stdlib.h>
+#include <inttypes.h>
 #include <errno.h>
 #include <string.h>
 #include <strings.h>
@@ -359,7 +360,7 @@ static struct mp3_data *mp3_open_internal (const char *file,
 		data->stream.sync = 0;
 		data->stream.error = MAD_ERROR_NONE;
 
-		if (io_seek(data->io_stream, 0, SEEK_SET) == (off_t)-1) {
+		if (io_seek(data->io_stream, 0, SEEK_SET) == -1) {
 			decoder_error (&data->error, ERROR_FATAL, 0, "seek failed");
 			io_close (data->io_stream);
 			mad_stream_finish (&data->stream);
@@ -399,8 +400,7 @@ static void *mp3_open_stream (struct io_stream *stream)
 	data->bitrate = -1;
 	data->io_stream = stream;
 	data->duration = -1;
-
-	data->size = (off_t)-1;
+	data->size = -1;
 
 	mad_stream_init (&data->stream);
 	mad_frame_init (&data->frame);
@@ -635,7 +635,7 @@ static int mp3_decode (void *void_data, char *buf, int buf_len,
 static int mp3_seek (void *void_data, int sec)
 {
 	struct mp3_data *data = (struct mp3_data *)void_data;
-	int new_position;
+	off_t new_position;
 
 	assert (sec >= 0);
 
@@ -648,7 +648,7 @@ static int mp3_seek (void *void_data, int sec)
 	new_position = ((double) sec /
 			(double) data->duration) * data->size;
 
-	debug ("Seeking to %d (byte %d)", sec, new_position);
+	debug ("Seeking to %d (byte %"PRId64")", sec, new_position);
 
 	if (new_position < 0)
 		new_position = 0;
@@ -656,7 +656,7 @@ static int mp3_seek (void *void_data, int sec)
 		return -1;
 
 	if (io_seek(data->io_stream, new_position, SEEK_SET) == -1) {
-		logit ("seek to %d failed", new_position);
+		logit ("seek to %"PRId64" failed", new_position);
 		return -1;
 	}
